upload

+++ Variable/Source03_13.cs
@@ -6,7 +6,7 @@ class MainClass
     {
         sbyte value1 = 64;
         sbyte value2 = 64;
-        int value3 = value1 + value2;
+        int value3 = (int)value1 + value2;
         Console.WriteLine(value3);
     }
 }
